Move markdownz story
